+++ frontend/src/components/App.tsx
@@ -1,16 +1,17 @@
 import React from 'react'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import useMediaQuery from '@mui/material/useMediaQuery'
-import Homepage from './pages/Homepage'
-import { ProgramDetails } from './pages/programDetails/ProgramDetails'
-import { WorkoutDetails } from './pages/workoutDetails'
-import { WorkoutHistory } from './pages/workoutHistory'
-import { AddProgram } from './pages/AddProgram'
-import { Planner, WorkoutList } from './pages/planner'
+import { Homepage } from './pages/homepage'
+import { ProgramDetails } from './pages/program-details/ProgramDetails'
+import { WorkoutDetails } from './pages/workout-details'
+import { WorkoutHistory } from './pages/workout-history'
+import { AddProgram } from './pages/add-program/AddProgram'
+import { Planner } from './pages/planner'
+import { WorkoutList } from './pages/planner/components'
 import GlobalStyle from '../styles/GlobalStyles'
-import NotFound from './pages/NotFound'
+import { NotFound } from './pages/not-found'
 import { ProtectedRoute } from './auth/components'
-import NotMobile from './pages/NotMobile'
+import { NotMobile } from './pages/not-mobile'
 import { Header } from './common'
 
 const App: React.FC = () => {
