+++ frontend/src/components/pages/add-program/index.ts
@@ -1,0 +1,3 @@
+import { AddProgram } from './AddProgram'
+
+export { AddProgram }
