+++ frontend/src/components/pages/homepage/index.ts
@@ -1,0 +1,3 @@
+import { Homepage } from './Homepage'
+
+export { Homepage }
