+++ frontend/src/components/pages/not-found/NotFound.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
-import { NavigationButton } from '../common'
+import { NavigationButton } from '../../common'
 import { Box, Typography } from '@mui/material'
 import { FiXCircle } from 'react-icons/fi'
 
-const NotFound: React.FC = () => {
+export const NotFound: React.FC = () => {
   return (
     <Box
       sx={{
@@ -23,5 +23,3 @@ const NotFound: React.FC = () => {
     </Box>
   )
 }
-
-export default NotFound
