+++ frontend/src/components/pages/not-found/index.ts
@@ -1,0 +1,3 @@
+import { NotFound } from './NotFound'
+
+export { NotFound }
