+++ frontend/src/components/pages/not-mobile/index.ts
@@ -1,0 +1,3 @@
+import { NotMobile } from './NotMobile'
+
+export { NotMobile }
