+++ frontend/src/components/pages/planner/components/ProgramList.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react'
-import type Program from '../../../types/Program'
+import type Program from '../../../../types/Program'
 import { Box } from '@mui/material'
-import { NavigationButton } from '../../common'
+import { NavigationButton } from '../../../common'
 import { useAuth0 } from '@auth0/auth0-react'
-import { fetchUserPrograms } from '../../../utils/fetchUserPrograms'
+import { fetchUserPrograms } from '../../../../utils/fetchUserPrograms'
 import CircularProgress from '@mui/material/CircularProgress'
 
 export const ProgramList: React.FC = () => {
