+++ frontend/src/components/pages/planner/components/WorkoutList.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Typography from '@mui/material/Typography'
 import Box from '@mui/material/Box'
-import { ProgramList } from '.'
+import { ProgramList } from './'
 
 export const WorkoutList: React.FC = () => {
   return (
