+++ frontend/src/components/pages/planner/components/index.ts
@@ -1,0 +1,4 @@
+import { ProgramList } from './ProgramList'
+import { WorkoutList } from './WorkoutList'
+
+export { ProgramList, WorkoutList }
