+++ frontend/src/components/pages/planner/index.ts
@@ -1,5 +1,3 @@
 import { Planner } from './Planner'
-import { ProgramList } from './ProgramList'
-import { WorkoutList } from './WorkoutList'
 
-export { Planner, ProgramList, WorkoutList }
+export { Planner }
