+++ frontend/src/components/pages/program-details/components/ConfirmDeleteDialog.tsx
