+++ frontend/src/components/pages/program-details/index.ts
@@ -1,0 +1,3 @@
+import { ProgramDetails } from './ProgramDetails'
+
+export { ProgramDetails }
