+++ frontend/src/components/pages/workout-details/components/index.ts
@@ -1,0 +1,3 @@
+import { WorkoutDetailsForm } from './WorkoutDetailsForm'
+
+export { WorkoutDetailsForm }
