+++ frontend/src/components/pages/workout-details/index.ts
@@ -1,0 +1,3 @@
+import { WorkoutDetails } from './WorkoutDetails'
+
+export { WorkoutDetails }
