+++ frontend/src/components/pages/workout-history/components/WorkoutHistoryTableRow.tsx
@@ -9,7 +9,7 @@ import {
   TableBody,
   Table
 } from '@mui/material'
-import type Workout from '../../../types/Workout'
+import type Workout from '../../../../types/Workout'
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp'
 import { FaCheck } from 'react-icons/fa'
