+++ frontend/src/components/pages/workout-history/components/index.ts
@@ -1,0 +1,3 @@
+import { WorkoutHistoryTableRow } from './WorkoutHistoryTableRow'
+
+export { WorkoutHistoryTableRow }
