+++ frontend/src/components/pages/workout-history/index.ts
@@ -1,0 +1,3 @@
+import { WorkoutHistory } from './WorkoutHistory'
+
+export { WorkoutHistory }
