--- frontend/src/components/pages/workoutDetails/index.ts
@@ -1,4 +1,0 @@
-import { WorkoutDetails } from './WorkoutDetails'
-import { WorkoutDetailsForm } from './WorkoutDetailsForm'
-
-export { WorkoutDetails, WorkoutDetailsForm }
