--- frontend/src/components/pages/workoutHistory/index.ts
@@ -1,4 +1,0 @@
-import { WorkoutHistory } from './WorkoutHistory'
-import { WorkoutHistoryTableRow } from './WorkoutHistoryTableRow'
-
-export { WorkoutHistory, WorkoutHistoryTableRow }
